+++ adaptor/typed/typedjournal/search.go
@@ -3,6 +3,7 @@ package typedjournal
 import (
 	"context"
 
+	"github.com/dogmatiq/persistencekit/adaptor/typed/typedmarshaler"
 	"github.com/dogmatiq/persistencekit/journal"
 )
 
@@ -11,21 +12,24 @@ import (
 // If the record is less than the datum, cmp is negative. If the record is
 // greater than the datum, cmp is positive. Otherwise, the record is considered
 // equal to the datum.
-type CompareFunc[R any] func(
+type CompareFunc[Record any] func(
 	ctx context.Context,
 	pos journal.Position,
-	rec R,
+	rec Record,
 ) (cmp int, err error)
 
 // BinarySearch performs a binary search of the journal to find the position of
 // the record for which cmp() returns zero.
-func BinarySearch[R any, M Marshaler[R]](
+func BinarySearch[
+	Record any,
+	Marshaler typedmarshaler.Marshaler[Record],
+](
 	ctx context.Context,
-	j Journal[R, M],
+	j Journal[Record, Marshaler],
 	begin, end journal.Position,
-	cmp CompareFunc[R],
-) (journal.Position, R, error) {
-	var rec R
+	cmp CompareFunc[Record],
+) (journal.Position, Record, error) {
+	var rec Record
 	pos, _, err := journal.BinarySearch(
 		ctx,
 		j.Journal,