+++ adaptor/typed/typedkv/store.go
@@ -3,15 +3,20 @@ package typedkv
 import (
 	"context"
 
+	"github.com/dogmatiq/persistencekit/adaptor/typed/typedmarshaler"
 	"github.com/dogmatiq/persistencekit/kv"
 )
 
 // Store is a collection of keyspaces that maps keys of type K to values of type
 // V.
-type Store[K, V any, KM Marshaler[K], VM Marshaler[V]] struct {
+type Store[
+	Key, Value any,
+	KeyMarshaler typedmarshaler.Marshaler[Key],
+	ValueMarshaler typedmarshaler.Marshaler[Value],
+] struct {
 	kv.Store
-	KeyMarshaler   KM
-	ValueMarshaler VM
+	KeyMarshaler   KeyMarshaler
+	ValueMarshaler ValueMarshaler
 }
 
 // Open returns the journal with the given name.