+++ CHANGELOG.md
@@ -10,6 +10,21 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## Unreleased
+
+### Changed
+
+- **[BC]** Moved `typedjournal` and `typedkv` packages into `adaptor` directory.
+
+### Added
+
+- Added `typedmarshaler` package.
+
+### Removed
+
+- **[BC]** Removed `typedjournal.Marshaler` and `typedkv.Marshaler`.
+  Use `typedmarshaler.Marshaler` instead.
+
 ## [0.3.0] - 2024-03-09
 
 ### Added